Moving most useful functions as library
