+++ include/io/prepare.hpp
@@ -3,7 +3,7 @@
 #include <fstream>
 #include <iostream>
 
-#include "../defaults.hpp"
+#include "defaults.hpp"
 #include "../context.hpp"
 
 std::string prepExperiment(Context* ctx);
