+++ lib/bitspin.hpp
@@ -1,0 +1,7 @@
+#include "defaults.hpp"
+
+double bool2spin(bool);
+double bool2spin(uint, int);
+
+bool bitFromEnd(uWord_t number, uint a);
+bool bitFromBeg(uWord_t number, uint a);
